requirements.txt
keras==2.11.0
numpy==1.24.2
opencv_python==4.7.0.68
Pillow==9.4.0
requests==2.28.1
streamlit==1.17.0
tensorflow==2.11.0
tensorflow_intel==2.11.0
